better practice changes
