+++ projectDiary.txt
@@ -1,0 +1,3 @@
+31-12-2024 -- rahul
+-- signup component is split into UI and interactivity (signUpForm) components
+-- Created styleClasses directory to get classnames (With styles ofc) for reuse
