+++ src/app/page.js
@@ -1,9 +1,5 @@
-import SignUp from "@/components/signUp";
+import SignUp from "@/components/signup/signUp";
 
 export default function Home() {
-  return (
-    <div>
-      Main page
-    </div>
-  );
+  return <div>Main page</div>;
 }
