+++ src/app/signup/page.js
@@ -1,14 +1,10 @@
-import SignUp from "@/components/signUp";
+import SignUp from "@/components/signup/signUp";
 import SignUpCardHolder from "@/components/custom-ui/signUpCardHolder";
 
-export default function SignupPage(){
-
-    return(
-        <div className="flex justify-center items-center min-h-screen bg-signup-bg relative">
-            <SignUpCardHolder>
-                {<SignUp/>}
-            </SignUpCardHolder>
-        </div>
-        
-    )
+export default function SignupPage() {
+  return (
+    <div className="flex justify-center items-center min-h-screen bg-signup-bg relative">
+      <SignUpCardHolder>{<SignUp />}</SignUpCardHolder>
+    </div>
+  );
 }
